requirements.txt
absl-py==0.9.0
aiodns==2.0.0
aiohttp==3.6.2
aiohttp-socks==0.3.9
astunparse==1.6.3
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.1
cachetools==4.1.0
cchardet==2.1.6
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cycler==0.10.0
elasticsearch==7.7.1
fake-useragent==0.1.11
gast==0.3.3
geographiclib==1.50
geopy==1.22.0
google-auth==1.17.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googletransx==2.4.2
grpcio==1.29.0
h5py==2.10.0
idna==2.9
idna-ssl==1.1.0
importlib-metadata==1.6.1
joblib==0.15.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.2.1
multidict==4.7.6
nltk==3.5
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.2.1
pandas==1.0.4
Pillow==7.1.2
protobuf==3.12.2
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==3.1.1
pycparser==2.20
pyparsing==2.4.7
PyQt5==5.15.0
PyQt5-sip==12.8.0
PySocks==1.7.1
python-crfsuite==0.9.7
python-dateutil==2.8.1
pytz==2020.1
pyvi==0.1
PyYAML==5.3.1
regex==2020.6.8
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.1
schedule==0.6.0
scikit-learn==0.23.1
scipy==1.4.1
six==1.15.0
sklearn-crfsuite==0.3.6
soupsieve==2.0.1
tabulate==0.8.7
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
threadpoolctl==2.1.0
tqdm==4.46.1
twint==2.1.20
typing==3.7.4.1
typing-extensions==3.7.4.2
Unidecode==1.1.1
urllib3==1.25.9
Werkzeug==1.0.1
wordcloud==1.7.0
wrapt==1.12.1
yarl==1.4.2
zipp==3.1.0
